requirements.txt
Flask
gunicorn
sqlalchemy
python-dotenv
psycopg2
requests
flask_sqlalchemy